+++ src/main/java/com/test/repository/RolesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolesRepository extends JpaRepository<Roles, Byte> {
+public interface RolesRepository extends JpaRepository<Roles, Short> {
     public Roles getByRoleName(String roleName);
 }
